Rename neovim config directory
