+++ nvim/coc-settings.json
