+++ nvim/options.lua
