+++ nvim/plugins.lua
